create ticket

+++ Controllers/ShowTicket.php
@@ -7,29 +7,24 @@ use Illuminate\Contracts\Container\BindingResolutionException;
 use Leantime\Core\Controller\Controller;
 use Symfony\Component\HttpFoundation\Response;
 use Leantime\Plugins\ShowTicket\Services\ShowTicket as ShowTicketService;
-use Leantime\Core\Language as LanguageCore;
 use Leantime\Core\UI\Template;
 use Leantime\Domain\Users\Services\Users as UserService;
 use Illuminate\Http\JsonResponse as JsonResponse;
 use Leantime\Domain\Sprints\Services\Sprints as SprintService;
 use Leantime\Domain\Tickets\Services\Tickets as TicketService;
-use Leantime\Domain\Tickets\Repositories\Tickets as TicketRepository;
 use Leantime\Domain\Files\Repositories\Files as FileRepository;
 use Leantime\Domain\Auth\Services\Auth as AuthService;
 use Leantime\Domain\Auth\Models\Roles;
-
 /**
  * ShowTicket controller.
  */
 class ShowTicket extends Controller
 {
     private ShowTicketService $showTicketService;
-    protected LanguageCore $language;
     protected Template $template;
     private UserService $userService;
     private SprintService $sprintService;
     private TicketService $ticketService;
-    private TicketRepository $ticketRepository;
     private FileRepository $filesRepo;
 
     /**
@@ -39,21 +34,17 @@ class ShowTicket extends Controller
      * @param TicketService     $ticketService
      * @param SprintService     $sprintService
      * @param UserService       $userService
-     * @param LanguageCore      $language
      * @param Template          $template
-     * @param TicketRepository  $ticketRepository
      * @param FileRepository    $filesRepo
      * @return void
      */
-    public function init(ShowTicketService $showTicketService, TicketService $ticketService, SprintService $sprintService, UserService $userService, LanguageCore $language, Template $template, TicketRepository $ticketRepository, FileRepository $filesRepo): void
+    public function init(ShowTicketService $showTicketService, TicketService $ticketService, SprintService $sprintService, UserService $userService, Template $template, FileRepository $filesRepo): void
     {
         $this->showTicketService = $showTicketService;
         $this->ticketService = $ticketService;
         $this->sprintService = $sprintService;
-        $this->language = $language;
         $this->userService = $userService;
         $this->template = $template;
-        $this->ticketRepository = $ticketRepository;
         $this->filesRepo = $filesRepo;
     }
 
@@ -66,6 +57,9 @@ class ShowTicket extends Controller
      */
     public function saveTicket(array $input): JsonResponse
     {
+        if (!AuthService::userIsAtLeast(Roles::$editor)) {
+            return  response()->json(["error"=>true]);
+        }
         $saveResult = $this->showTicketService->saveTicket($input['id'], $input['key'], $input['value']);
         return response()->json(['ticket' => $saveResult]);
     }
@@ -79,12 +73,15 @@ class ShowTicket extends Controller
      */
     public function deleteTicket(array $input): JsonResponse
     {
+        if (!AuthService::userIsAtLeast(Roles::$editor)) {
+            return  response()->json(["error"=>true]);
+        }
         $deleteResult = $this->showTicketService->deleteTicket($input['id']);
         return response()->json(['ticket' => $deleteResult]);
     }
 
     /**
-     * Saves ticket headline.
+     * Get tags
      *
      * @param string[] $input The input for getting tags:
      *                     - 'id': The project id.
@@ -93,27 +90,12 @@ class ShowTicket extends Controller
      */
     public function getTags(array $input): JsonResponse
     {
-        $tags = [];
-        $ticketTags = $this->ticketRepository->getTags($input['projectId']);
-        $tags = $this->explodeAndMergeTags($ticketTags, $tags);
-        $uniqueTags = array_unique($tags);
-        return response()->json(['tags' => $uniqueTags]);
-    }
-
-    // phpcs:disable
-    /** @phpstan-ignore-next-line */
-    private function explodeAndMergeTags(array $dbTagValues, array $mergeInto): array
-    {
-        foreach ($dbTagValues as $tagGroup) {
-            if (isset($tagGroup['tags']) && $tagGroup['tags'] != null) {
-                $tagArray = explode(',', $tagGroup['tags']);
-                $mergeInto = array_merge($tagArray, $mergeInto);
-            }
+        if (!AuthService::userIsAtLeast(Roles::$editor)) {
+            return  response()->json(["error"=>true]);
         }
-
-        return $mergeInto;
+        $tags = $this->showTicketService->getTags((int)$input['projectId']);
+        return response()->json(['tags' => $tags]);
     }
-    // phpcs:enable
 
     /**
      * @return Response
@@ -122,7 +104,7 @@ class ShowTicket extends Controller
     public function post(): Response
     {
         if (!AuthService::userIsAtLeast(Roles::$editor)) {
-            return $this->tpl->displayJson(['Error' => 'Not Authorized'], 403);
+            return  response()->json(["error"=>true]);
         }
         $redirectUrl = BASE_URL . '/ShowTicket/ShowTicket';
         if (isset($_POST['ticket-id'])) {
@@ -131,7 +113,6 @@ class ShowTicket extends Controller
 
         return Frontcontroller::redirect($redirectUrl);
     }
-
 
     /**
      * get
@@ -142,29 +123,46 @@ class ShowTicket extends Controller
      * @throws BindingResolutionException
      */
     public function get(): Response
-    {
+    {   
+        if (!AuthService::userIsAtLeast(Roles::$editor)) {
+            return  response()->json(["error"=>true]);
+        }
+        $this->template->assign('ticketExists', false);
         $ticket = null;
-        if (array_key_exists('ticketId', $_GET)) {
+        if (array_key_exists('ticketId', $_GET) && $_GET['ticketId'] !== NULL) {
             $ticket = $this->showTicketService->getTicket($_GET['ticketId']);
+            $this->template->assign('ticketIdFromUrl', $_GET['ticketId']);
 
             if ($ticket) {
                 $statusLabels = $this->showTicketService->getStatusLabels($ticket->projectId);
-                $priorityLabels = $this->showTicketService->getPriorityLabels();
-                $this->template->assign('ticket', $ticket);
                 $this->template->assign('statusLabels', $statusLabels);
-                $this->template->assign('priorityLabels', $priorityLabels);
-                $this->tpl->assign('sprints', $this->sprintService->getAllSprints($ticket->projectId));
-                $this->tpl->assign('allUsers', $this->userService->getAll());
-                $this->tpl->assign('relatedTickets', $this->ticketService->getAllPossibleParents($ticket));
-                $milestones = $this->ticketService->getAllMilestones(['sprint' => '', 'type' => 'milestone', 'currentProject' => $ticket->projectId]);
-                $this->tpl->assign('milestones', $milestones);
-                $this->tpl->assign('files', $this->filesRepo->getFilesByModule('ticket', $ticket->id));
-                $this->tpl->assign('subtasks', $this->showTicketService->getAllSubtasks($ticket->id));
-            }
 
-            $this->tpl->assign('ticketIdFromUrl', $_GET['ticketId']);
+                $priorityLabels = $this->showTicketService->getPriorityLabels();
+                $this->template->assign('priorityLabels', $priorityLabels);
+
+                $sprints = $this->sprintService->getAllSprints($ticket->projectId);
+                $this->template->assign('sprints', $sprints);
+
+                $users = $this->userService->getAll();
+                $this->template->assign('allUsers', $users);
+
+                $milestones = $this->ticketService->getAllMilestones(['sprint' => '', 'type' => 'milestone', 'currentProject' => $ticket->projectId]);
+                $this->template->assign('milestones', $milestones);
+
+                $relatedTickets = $this->ticketService->getAllPossibleParents($ticket);
+                $this->template->assign('relatedTickets', $relatedTickets);
+
+                $files = $this->filesRepo->getFilesByModule('ticket', $ticket->id);
+                $this->template->assign('files', $files);
+
+                $subtasks = $this->showTicketService->getAllSubtasks($ticket->id);
+                $this->template->assign('subtasks', $subtasks);
+
+                $this->template->assign('ticketExists', true);
+                $this->template->assign('ticket', $ticket);
+            }
         }
-        // Ticket assigned to the template
+        
         return $this->template->display('ShowTicket.showTicket');
     }
 }
